+++ src/types/forms/UpdateUserFormData.ts
@@ -1,0 +1,13 @@
+export type UpdateUserFormData = {
+  name: { first: string; middle?: string; last: string };
+  phone: string;
+  image: { url: string; alt: string };
+  address: {
+    state?: string;
+    country: string;
+    city: string;
+    street: string;
+    houseNumber: number;
+    zip: number;
+  };
+};